fix: rename agentskills-cli distribution to agentskills-tools
PyPI rejected agentskills-cli: its name-similarity check ultranormalizes names before comparing, and the unrelated agent-skills-cli project already collides. The v0.4.0 publish run failed on that package with a 400, which also skipped agentskills-testing and the GitHub Release.

Renames the distribution to agentskills-tools and the module to agentskills_tools, moves the package to packages/tools/, and updates the workspace config, coverage floors, release scripts, workflows, validate action, docs and mkdocs nav to match. The SDK logger namespace for these modules becomes agentskills.tools.*.
