+++ .github/workflows/docs.yml
@@ -45,7 +45,7 @@ jobs:
             ./packages/integrations/agentskills-langchain \
             ./packages/integrations/agentskills-agentframework \
             ./packages/integrations/agentskills-mcp-server \
-            ./packages/cli/agentskills-cli \
+            ./packages/tools/agentskills-tools \
             ./packages/testing/agentskills-testing
 
       - name: Build site
@@ -78,7 +78,7 @@ jobs:
             ./packages/integrations/agentskills-langchain \
             ./packages/integrations/agentskills-agentframework \
             ./packages/integrations/agentskills-mcp-server \
-            ./packages/cli/agentskills-cli \
+            ./packages/tools/agentskills-tools \
             ./packages/testing/agentskills-testing
 
       - name: Publish docs for release tag
