+++ .github/workflows/publish.yml
@@ -81,7 +81,7 @@ jobs:
             packages/integrations/agentskills-langchain \
             packages/integrations/agentskills-agentframework \
             packages/integrations/agentskills-mcp-server \
-            packages/cli/agentskills-cli \
+            packages/tools/agentskills-tools \
             packages/testing/agentskills-testing
           do
             name="$(basename "$pkg")"
@@ -168,10 +168,10 @@ jobs:
           attestations: true
           skip-existing: true
 
-      - name: agentskills-cli
+      - name: agentskills-tools
         uses: pypa/gh-action-pypi-publish@dc37677b2e1c63e2034f94d8a5b11f265b73ba33 # v1.14.2
         with:
-          packages-dir: dist/agentskills-cli
+          packages-dir: dist/agentskills-tools
           attestations: true
           skip-existing: true
