+++ docs/issues/v0.4.md
@@ -23,7 +23,7 @@ skill means copying an existing folder and hoping the frontmatter is right.
 
 ### Proposal
 
-A new `agentskills-cli` package exposing an `agentskills` entry point.
+A new `agentskills-tools` package exposing an `agentskills` entry point.
 
 | Command | Behaviour |
 |---|---|
@@ -53,7 +53,7 @@ integration.
 ### Implementation notes
 
 **Both open questions were answered the same way, for the same reason.** A separate
-`agentskills-cli` package rather than `agentskills-core[cli]`, and stdlib `argparse` rather
+`agentskills-tools` package rather than `agentskills-core[cli]`, and stdlib `argparse` rather
 than `click` or `typer`. Item 2 installs this CLI into other people's CI, so every dependency
 it carries is a supply-chain edge those repositories inherit for the privilege of checking a
 markdown file. `argparse` costs some polish in `--help`; that is the whole price.
@@ -103,9 +103,18 @@ rather than wrong.
 last one fails silently: everything else errors, but a package missing from `publish.yml`
 simply never ships. Recorded in `.github/copilot-instructions.md`.
 
-**`agentskills-cli` is a new PyPI project**, so it needs a pending trusted publisher registered
+**`agentskills-tools` is a new PyPI project**, so it needs a pending trusted publisher registered
 before the v0.4.0 tag — and the pending-publisher uniqueness limit documented in item 12 means
 it cannot be registered alongside another pending one.
+
+**The distribution is named `agentskills-tools`, not `agentskills-cli`.** PyPI rejected the
+latter: its name-similarity check ultranormalizes a name (lowercase, strip `-`/`_`/`.`, map
+`l`/`i` to `1` and `o` to `0`) before comparing, and the unrelated `agent-skills-cli` project
+already collides under that rule. The rename landed after the v0.4.0 tag had already published
+seven of the nine distributions, so the tag was recreated on the fix commit and the already-live
+packages were skipped on re-upload. Any new distribution name must be checked against
+`https://pypi.org/pypi/<name>/json` (404 means free) *and* against its hyphenated variants
+before a pending publisher is registered.
 
 ---
 
@@ -185,7 +194,7 @@ Buildkite all have annotation formats. The JSON output is the honest answer for 
 has written that adapter, so today the pull request experience is GitHub-only.
 
 **The action cannot be tagged until the CLI is on PyPI.** `pratikxpanda/agentskills-sdk/actions/validate@v1`
-needs a `v1` tag, and the action installs `agentskills-cli`, which does not exist on the index
+needs a `v1` tag, and the action installs `agentskills-tools`, which does not exist on the index
 until v0.4.0 ships. The README snippet is correct and will not work until then.
 
 **Nothing verifies the action's failure path in CI.** The dogfood job proves a valid skill
@@ -311,7 +320,7 @@ safe to share. 26s → 0.4s.
   obvious: a hand-written minimal provider inside core's test tree, probably.
 - **`agentskills-testing` is a new PyPI project.** It needs a pending trusted publisher registered
   before the v0.4.0 tag, and only one pending publisher can exist at a time across identically
-  claimed distributions — so it and `agentskills-cli` must be bootstrapped in sequence.
+  claimed distributions — so it and `agentskills-tools` must be bootstrapped in sequence.
 - **The conformance fixture contract is a compatibility surface.** Adding an assertion that needs
   new fixture content is a breaking change for anyone already using the suite. Worth an explicit
   policy before 1.0.
@@ -477,7 +486,7 @@ interpolated into URLs.
   generalise it to a provider-level concurrency setting or leave registration sequential
   permanently and say so; the current state is a coin flip waiting to be re-flipped.
 - **The "immediate subdirectory containing a `SKILL.md`" rule now exists twice**, in
-  `agentskills_cli.discovery` and in `LocalFileSystemSkillProvider._discover_sync`. The CLI was
+  `agentskills_tools.discovery` and in `LocalFileSystemSkillProvider._discover_sync`. The CLI was
   deliberately left alone: its `discover()` returns `SkillLocation(skill_id, path)` objects that
   the provider API cannot supply, and `serve.build_registry` may be handed a single-skill target
   whose root is the parent directory, where `register_all` would wrongly register the siblings.
@@ -618,7 +627,7 @@ and all four filter arguments.
 > **Status: implemented.** All three open questions were resolved toward the dependency-light
 > answer: a one-method protocol resolved from a dotted path, `repeat` plus `threshold` per case,
 > and an on-disk cache keyed by what actually determined each completion. The harness landed in
-> `agentskills-cli` rather than `agentskills-testing`; the reasoning is below.
+> `agentskills-tools` rather than `agentskills-testing`; the reasoning is below.
 
 **Labels:** `theme:agent-effectiveness`, `type:feature`, `package:cli`
 
@@ -671,7 +680,7 @@ declared judge model and should report agreement, since an LLM judge is itself u
 
 ### Implementation notes
 
-**The harness is in `agentskills-cli`, not `agentskills-testing`.** The `package:testing` label
+**The harness is in `agentskills-tools`, not `agentskills-testing`.** The `package:testing` label
 was a guess made before anyone tried to build it, and the acceptance criteria contradict it: a
 thing that must *never* run in the default `pytest` run is a command, not a pytest kit. Putting
 it in `agentskills-testing` would also have forced `agentskills validate` — which must keep
@@ -733,7 +742,7 @@ wrong default when the failure mode is money.
   contract into an agent-loop contract that every adapter author then has to implement. Worth
   doing when there is a caller for it, and worth doing as a separate protocol that
   `EvalModel` can opt into rather than a widening of `complete()`.
-- **`agentskills-cli` imports `yaml` without declaring `pyyaml`.** It works today only because
+- **`agentskills-tools` imports `yaml` without declaring `pyyaml`.** It works today only because
   `agentskills-core` pulls it in transitively. `evalspec.py` adds a second such import and makes
   the latent packaging bug more load-bearing. It could not be fixed in this change: adding a
   dependency requires `poetry lock`, and `poetry` currently cannot reach `files.pythonhosted.org`
