--- packages/cli/agentskills-cli/agentskills_cli/__main__.py
@@ -1,10 +1,0 @@
-"""Entry point for ``python -m agentskills_cli`` and the ``agentskills`` script."""
-
-from __future__ import annotations
-
-import sys
-
-from agentskills_cli.cli import main
-
-if __name__ == "__main__":
-    sys.exit(main())
