+++ scripts/dev.py
@@ -35,7 +35,7 @@ COVERAGE_FLOORS: dict[str, int] = {
     "agentskills_langchain": 100,
     "agentskills_agentframework": 100,
     "agentskills_mcp_server": 91,
-    "agentskills_cli": 99,
+    "agentskills_tools": 99,
     "agentskills_testing": 100,
 }
